requirements.txt
kani[openai]
gradio
openai
python-dotenv